pyproject.toml
[tool.poetry]
name = "tableau-api"
version = "0.1.0"
description = "Tableau API"
authors = [
"Mayra Valdes <4valdes@gmail.com>",
]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.11"
tableauserverclient = "^0.34"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
autopep8 = "^2.0.1"
flake8 = "^6.0.0"
pylint = "^2.16.1"
isort = "^5.12.0"
pytest-mock = "^3.10.0"
pre-commit = "^3.0.4"
bandit = "^1.7.4"
toml = "^0.10.2"
black = "^23.1.0"
syrupy = "^4.0.1"

[tool.pytest.ini_options]
testpaths = "tests"
filterwarnings = [
    "error",
    "ignore::DeprecationWarning"
]

[tool.bandit]
exclude_dirs = [".venv",]

[tool.bandit.assert_used]
skips = ['test_*.py', '*/test_*.py']


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"